requirements.txt
boto3==1.13.24
celery[sqs]==4.4.4
Django==3.0.8
django-celery-beat==2.0.0
django-compressor==2.4
django-extensions==2.2.9
django-localflavor==3.0.1
django-nose==1.4.6
django-object-actions==2.0.0
django-solo==1.1.3
django-stubs==1.5.0
django-wkhtmltopdf==3.3.0
future==0.18.2
googlemaps==4.4.1
gunicorn==20.0.4
kombu==4.6.10
localflavor==1.8
matplotlib==3.2.1
mypy==0.780
numpy==1.18.5
openpyxl==3.0.3
paramiko==2.7.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycurl==7.43.0.5
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
restless==2.1.1
sendgrid==6.3.1
sendgrid-django==4.2.0
twilio==6.41.0
git+ssh://git@github.com/pronym-inc/pronym-django-api.git@v2.0#egg=pronym_api
git+ssh://git@github.com/pronym-inc/stamps.git@v1.0#egg=stamps